+++ packages/f-header/CHANGELOG.md
@@ -4,6 +4,51 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v2.0.0-beta-21
+------------------------------
+*October 10, 2019*
+
+### Changed
+ - The way data is passed around in the order count logic
+ - Version number
+
+### Added
+ - Tests for order count logic
+
+
+v2.0.0-beta-20
+------------------------------
+*October 9, 2019*
+
+### Changed
+ - Prop default and value for `errorLog`
+
+
+v2.0.0-beta-19
+------------------------------
+*October 8, 2019*
+
+### Changed
+ - New version of `getWindowHeight` to `getWindowWidth`
+ - Italian tenant that was wrong
+
+
+v2.0.0-beta-18
+------------------------------
+*October 7, 2019*
+
+### Added
+ - New version of `f-services`
+
+
+v2.0.0-beta-17
+------------------------------
+*October 3, 2019*
+
+### Added
+ - New version of `f-services`
+
+
 v2.0.0-beta-16
 ------------------------------
 *October 1, 2019*